save dev state — refactor work.rs
